chore(assistant): move switch services for stt, llm and tts from `main.py` to their respective service files for the sake of better organization
