+++ .github/workflows/integration_delivery.yml
@@ -454,12 +454,9 @@ jobs:
           PKR_VAR_target_image_size:
             ${{ steps.generate-image-url.outputs.image_size}}
         run: |
-          ls /build/dist
           /entrypoint.sh init scripts/packer/image.pkr.hcl
           SETUP_QEMU=false /entrypoint.sh build scripts/packer/image.pkr.hcl
-          ls -lh
           mv image.img /build
-          ls -lh /build
 
       - name: Fill Free Space with Zeros
         run: |
