+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 - feat(core): add a mechanism for services to provide environment variables for their binaries
 - refactor(assistant): use the new secret querying service to query google and openai credentials/api-key
 - chore: run lint and typecheck for all sub-projects, fix lint issues in rpc and assistant sub-projects
+- chore(assistant): move switch services for stt, llm and tts from `main.py` to their respective service files for the sake of better organization
 
 ## Version 1.4.0
