+++ ubo_app/services/090-assistant/ubo-service/ubo_assistant/switch.py
@@ -86,8 +86,7 @@ class UboSwitchService(AIService, Generic[T]):
         if isinstance(frame, SystemFrame):
             await super().process_frame(frame, direction)
         if not self.selected_service:
-            msg = 'No service is selected'
-            raise ValueError(msg)
+            return
         await self.selected_service.process_frame(frame, direction)
 
     async def setup(self, setup: FrameProcessorSetup) -> None:
